requirements.txt
anyio==4.4.0
certifi==2024.8.30
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
numpy==2.1.0
pandas==2.2.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.5
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
